requirements.txt
flask==3.0.0
openai>=1.50.0
python-dotenv==1.0.0
pypdf==4.1.0
python-docx==1.1.0
